Visuelle justeringer på oversikten
