+++ frontend/mulighetsrommet-veileder-flate/src/components/oversikt/Tilgjengelighetsstatus.module.scss
