+++ frontend/mulighetsrommet-veileder-flate/src/components/oversikt/Tilgjengelighetsstatus.tsx
