+++ frontend/mulighetsrommet-veileder-flate/src/components/oversikt/Tiltaksgjennomforingsoversikt.module.scss
@@ -1,0 +1,16 @@
+.gjennomforinger {
+  margin: 0;
+  padding: 0;
+}
+
+.filter_loader {
+  display: block;
+  margin: 0 auto;
+}
+
+.under_oversikt {
+  display: flex;
+  justify-content: space-between;
+  margin-top: 1rem;
+  margin-bottom: 2rem;
+}
