+++ frontend/mulighetsrommet-veileder-flate/src/components/sidemeny/Sidemenydetaljer.module.scss
@@ -1,7 +1,7 @@
 .rad {
   display: flex;
   justify-content: space-between;
-  margin-bottom: 1rem;
+  margin-bottom: 1.5rem;
   column-gap: 1rem;
   &:last-child {
     margin-bottom: 0;
