--- frontend/mulighetsrommet-veileder-flate/src/components/tabell/Tabell.module.scss
@@ -1,56 +1,0 @@
-.filter_loader {
-  display: block;
-  margin: 0 auto;
-}
-
-.tabell_wrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  :global(.kopiknapp:hover) {
-    background-color: transparent;
-  }
-}
-
-.tabell {
-  &_tiltaksnavn {
-    font-weight: bold;
-  }
-
-  &_tiltaksnummer {
-    padding-right: 1.5rem;
-
-    @media (max-width: 1400px) {
-      grid-template-columns: 40% auto;
-    }
-  }
-
-  &_tiltaksnavn {
-    width: 25%;
-  }
-}
-
-.under_tabell {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 1rem;
-  margin-bottom: 2rem;
-}
-
-:global(.mulighetsrommet-veileder-flate) {
-  :global(.navds-table) {
-    border: 1px solid var(--navds-semantic-color-border-muted);
-
-    :global(.navds-table__sort-button) {
-      height: 4rem;
-    }
-  }
-
-  :global(.navds-table__header),
-  :global(.navds-table--zebra-stripes),
-  :global(.navds-table__body),
-  :global(.navds-table__row:nth-of-type(even):not(:hover):not(.navds-table__row--selected)) {
-    background-color: var(--navds-semantic-color-canvas-background-light);
-  }
-}
